refactor: moving types around and renaming things
