+++ src/Gen3Hex.Core/Gen3Hex.Core.csproj
@@ -43,12 +43,16 @@
     <Compile Include="..\..\artifacts\$(AssemblyName)\codegen\StubCommand.cs" />
     <Compile Include="..\..\artifacts\$(AssemblyName)\codegen\StubDisposable.cs" />
     <Compile Include="ICommandExtensions.cs" />
-    <Compile Include="Models\ArrayRun.cs" />
-    <Compile Include="Models\DeltaModel.cs" />
+    <Compile Include="Models\Runs\ArrayRun.cs" />
+    <Compile Include="Models\AutoSearchModel.cs" />
     <Compile Include="Models\ErrorInfo.cs" />
-    <Compile Include="Models\FormattedRun.cs" />
-    <Compile Include="Models\DataFormatModel.cs" />
+    <Compile Include="Models\IDataModel.cs" />
+    <Compile Include="Models\Runs\IFormattedRun.cs" />
+    <Compile Include="Models\ModelDelta.cs" />
+    <Compile Include="Models\Runs\PCSRun.cs" />
     <Compile Include="Models\PCSString.cs" />
+    <Compile Include="Models\Runs\PointerRun.cs" />
+    <Compile Include="Models\PokemonModel.cs" />
     <Compile Include="Models\SearchByte.cs" />
     <Compile Include="Models\StoredMetadata.cs" />
     <Compile Include="NumericExtensions.cs" />
@@ -60,11 +64,14 @@
     <Compile Include="ViewModels\ChildViewPort.cs" />
     <Compile Include="ViewModels\DataFormats.cs" />
     <Compile Include="ViewModels\EditorViewModel.cs" />
+    <Compile Include="ViewModels\GotoControlViewModel.cs" />
     <Compile Include="ViewModels\HexElement.cs" />
     <Compile Include="ViewModels\IChildViewPort.cs" />
     <Compile Include="ViewModels\ITabContent.cs" />
-    <Compile Include="ViewModels\IToolViewModel.cs" />
+    <Compile Include="ViewModels\Tools\IToolTrayViewModel.cs" />
+    <Compile Include="ViewModels\Tools\IToolViewModel.cs" />
     <Compile Include="ViewModels\IViewPort.cs" />
+    <Compile Include="ViewModels\Tools\PCSTool.cs" />
     <Compile Include="ViewModels\ScrollRegion.cs" />
     <Compile Include="ViewModels\SearchResultsViewPort.cs" />
     <Compile Include="ViewModels\Selection.cs" />
