+++ src/Gen3Hex.Core/Models/PCSString.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Globalization;
-using System.Linq;
 using System.Text;
 
 namespace HavenSoft.Gen3Hex.Core.Models {
