+++ src/Gen3Hex.Core/ViewModels/Tools/IToolViewModel.cs
@@ -1,0 +1,16 @@
+﻿using System.ComponentModel;
+
+namespace HavenSoft.Gen3Hex.Core.ViewModels.Tools {
+   public interface IToolViewModel : INotifyPropertyChanged {
+      string Name { get; }
+   }
+
+   public class FillerTool : IToolViewModel {
+      public string Name { get; }
+      public FillerTool(string name) { Name = name; }
+
+#pragma warning disable 0067 // it's ok if events are never used
+      public event PropertyChangedEventHandler PropertyChanged;
+#pragma warning restore 0067
+   }
+}
